// src/minimact-spatial/src/types.ts
export interface AbsoluteBounds {
  top: number;
  left: number;
  width: number;
  height: number;
  right: number;
  bottom: number;
}